# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TD Investment System API
  version: 1.0.0
  description: >-
    API specification for TD Investment and Micro Credit backend services
    (Auth, Customers, Documents, Loans, Risk, Approvals).
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /api/customers:
    get:
      summary: List customers
      tags: [Customers]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
    post:
      summary: Create a new customer
      tags: [Customers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
...
  /api/documents:
    post:
      summary: Upload a document for a customer
      tags: [Documents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [customer_id, file]
              properties:
                customer_id:
                  type: integer
                file:
                  type: string
                  format: binary
                doc_type:
                  type: string
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        created_at:
          type: string
          format: date-time
    CustomerCreate:
      type: object
      required: [first_name, last_name, email]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    CustomerList:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
    Document:
      type: object
      properties:
        id:
          type: integer
        customer_id:
          type: integer
        doc_type:
          type: string
        filename:
          type: string
        url:
          type: string
        uploaded_at:
          type: string
          format: date-time
...